dernière version de la polyclinique

+++ module/Facturation/src/Facturation/Model/AdmissionTable.php
@@ -44,7 +44,7 @@ class AdmissionTable {
 				'a.date_cons' => $date
 		) );
 		
-		$select->order ( 'id_admission ASC' );
+		$select->order ( 'id_admission DESC' );
 		$stat = $sql->prepareStatementForSqlObject ( $select );
 		$result = $stat->execute ();
 		return $result;
@@ -71,8 +71,62 @@ class AdmissionTable {
 		$this->tableGateway->insert($donnees);
 	}
 	
-	public function deleteAdmissionPatient($id){
-		$this->tableGateway->delete(array('id_admission'=> $id));
+	/*
+	 * Recupérer la liste des patients admis et déjà consultés pour aujourd'hui
+	 */
+	public function getPatientAdmisCons(){
+		$today = new \DateTime ( 'now' );
+		$date = $today->format ( 'Y-m-d' );
+		
+		$adapter = $this->tableGateway->getAdapter ();
+		$sql = new Sql ( $adapter );
+		$select = $sql->select ( 'consultation' );
+		$select->columns ( array (
+				'ID_PATIENT'
+		) );
+		$select->where ( array (
+				'DATEONLY' => $date,
+		) );
+		
+		$result = $sql->prepareStatementForSqlObject ( $select )->execute ();
+		$tab = array();
+		foreach ($result as $res) {
+			$tab[] = $res['ID_PATIENT'];
+		}
+		
+		return $tab;
+	}
+	
+	/*
+	 * Fonction qui vérifie est ce que le patient n'est pas déja consulté
+	 */
+	public function verifierPatientConsulter($idPatient, $idService){
+		$today = new \DateTime ( 'now' );
+		$date = $today->format ( 'Y-m-d' );
+		
+		$adapter = $this->tableGateway->getAdapter ();
+		$sql = new Sql ( $adapter );
+		$select = $sql->select ( 'consultation' );
+		$select->columns ( array (
+				'ID_PATIENT'
+		) );
+		$select->where ( array (
+				'DATEONLY' => $date,
+				'ID_SERVICE' => $idService,
+				'ID_PATIENT' => $idPatient,
+		) );
+		
+		return $sql->prepareStatementForSqlObject ( $select )->execute ()->current();
+	}
+	
+	public function deleteAdmissionPatient($id, $idPatient, $idService){
+		if($this->verifierPatientConsulter($idPatient, $idService)){
+		    return 1;
+		} else {
+			$this->tableGateway->delete(array('id_admission'=> $id));
+			return 0;
+		}
+
 	}
 	
 	public function getPatientAdmis($id){
@@ -86,4 +140,57 @@ class AdmissionTable {
 		}
 		return $row;
 	}
+	
+	
+	public function getLastAdmission() {
+		$db = $this->tableGateway->getAdapter();
+		$sql = new Sql($db);
+		$sQuery = $sql->select('admission')
+		->order('id_admission DESC');
+		$requete = $sql->prepareStatementForSqlObject($sQuery);
+		return $requete->execute()->current();
+	}
+	
+	//Ajouter la consultation dans la table << consultation >> pour permettre au medecin de pouvoir lui même ajouter les constantes
+	//Ajouter la consultation dans la table << consultation >> pour permettre au medecin de pouvoir lui même ajouter les constantes
+	public function addConsultation($values , $IdDuService){
+		$today = new \DateTime ( 'now' );
+		$date = $today->format ( 'Y-m-d H:i:s' );
+		$dateOnly = $today->format ( 'Y-m-d' );
+		
+		$db = $this->tableGateway->getAdapter();
+		$this->tableGateway->getAdapter()->getDriver()->getConnection()->beginTransaction();
+		try {
+	
+			$dataconsultation = array(
+					'ID_CONS'=> $values->get ( "id_cons" )->getValue (),
+					'ID_PATIENT'=> $values->get ( "id_patient" )->getValue (),
+					'DATE'=> $date,
+ 					'DATEONLY' => $dateOnly,
+					'HEURECONS' => $values->get ( "heure_cons" )->getValue (),
+					'ID_SERVICE' => $IdDuService
+			);
+			
+			$sql = new Sql($db);
+			$sQuery = $sql->insert()
+			->into('consultation')
+			->values($dataconsultation);
+			$sql->prepareStatementForSqlObject($sQuery)->execute();
+	
+			$this->tableGateway->getAdapter()->getDriver()->getConnection()->commit();
+		} catch (\Exception $e) {
+			$this->tableGateway->getAdapter()->getDriver()->getConnection()->rollback();
+		}
+	}
+	
+	public function addConsultationEffective($id_cons){
+		$db = $this->tableGateway->getAdapter();
+		$sql = new Sql($db);
+		$sQuery = $sql->insert()
+		->into('consultation_effective')
+		->values(array('ID_CONS' => $id_cons));
+		$requete = $sql->prepareStatementForSqlObject($sQuery);
+		$requete->execute();
+	}
+	
 }
